requirements.txt
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Deprecated==1.2.7
idna==2.8
importlib-metadata==1.4.0
more-itertools==8.1.0
packaging==20.0
pluggy==0.13.1
py==1.8.1
PyGithub==1.45
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.2
requests==2.22.0
six==1.13.0
urllib3==1.25.7
wcwidth==0.1.8
wrapt==1.11.2
zipp==0.6.0
